Create site_lang.php
This is the (first version of the) Greek translation. I hope everything is ok. Check it out and please let me know when it is online so we could check it again in the website. If there are any apparent mistakes do not hesitate to contact us. Thank you very much